LayoutTests/plugins/npruntime/plugin-scriptable-object-invoke-default-expected.txt


This page tests invoking as functions JavaScript objects that are provided by plugins.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof(pluginWithInvokeDefault) is 'function'
PASS pluginWithInvokeDefault() is 1
PASS typeof(pluginWithoutInvokeDefault) is 'object'
PASS pluginWithoutInvokeDefault() threw exception TypeError: pluginWithoutInvokeDefault is not a function. (In 'pluginWithoutInvokeDefault()', 'pluginWithoutInvokeDefault' is an instance of HTMLEmbedElement).
